Replace curio for asyncio to accomodate aioice and aioquic
As the project grows to utilize P2P technologies, it has become increasingly clear that the project would have to rely on external libraries for handling some of the communication.

At the moment, curio lacks in relevant libraries to ease the use of protocols such as UDT, TURN, UPnP, PCP, ICE, QUIC and other UDP-based protocols, which are the surest way to achieve true P2P communication.

In moving away to the more widely-accepted asyncio library, the hope for the project is to allow seamless integration with aioice and aioquic. With these libraries, it should be relatively straight-forward to setup much of the communication necessary for P2P.

Another great advantage of asyncio is that it is so much more popular than curio, and is substantially more well supported by other libraries. Using prebuilt solutions based on asyncio is likely to increase the extensibility of the project substantially.