Add basic unit-tests for the project
After many deliberations, I have decided that it is best to comprehensively test the project before moving on with it. Creating a good test-suite will provide my project with certainty that commits aren't breaking any existing functionality, and will also help direct the project.

While this is but the first "test" test, it will serve as precedent for how it should be done. Using the famous pytest library.