Review the tests
It seems the tests have revealed a critical flaw in the design of the filesystem - that the interaction with the filesystem itself is troublesome for testing purposes, as it frustratingly requires to create files in the system just to create the tree representation.

It is therefore advised that the firesystem module be refactored to exclude any responsibility over filereading operations, which should be delegated to a different component. A file can be trivially represented as a file-name and raw data, and a directory as a directory-name and an **unordered** sequence of File | Directory.

Generating them would therefore become more trivial and straight-forward, and a separation of concerns be maintained.