Remove unnecessary tests
The tests don't need to verify builtin features of the language, or features of external dependencies. The tests removed are testing builtin functionality that is unnecessary.

Further question arise about the utility of fixtures, and what does it mean to create a fixture of a file with the absence of an actual file.