Remove JSON features from filesystem
Also add tests for classes using typing.Protocol to clarify the API specs for filesystem.
It turns out JSON is unnecessary, as the minimum requirements to satisfy tests only require converting to dictionary.

The functionality has been used somewhat in messages, but not extensively. It is unlikely this would be useful outside of networking-related code.

The dictionary form of File and Directory are a simplified, hashable, and serializable version of themselve. Dataclasses do not have these properties.