Rework file searching
Search is now a method of Directory type, not File.
File "searching" has been moved to __contains__, which clarifies the operation significantly.

In the context of a file, search would imply an operation on the file's contents, not the name. In that sense "searching" implies working on the content of an object, not on its metadata.