Create a new tcp_network module + test
The purpose of the new tcp_network module is to compete with protocol/network, server/network and client/network.
Unlike the earlier network modules, the new module can run in both server and client modes, and is fully tested. It does not depend on protocol/messages, and delegates the responsibility of encoding and decoding the messages to a different unplanned module.

By disentangling messages and netcode, the network code becomes easier to replace with alternative implementations using UDP or other network protocols, which is definitely planned.